Audit fixes: safety, security, docs consistency across firmware, ROS2, PWA
Firmware (13 fixes):
- Non-blocking WiFi reconnect in loop (was blocking 15s, freezing motors)
- Motor ramp int8_t overflow fix (prevented sudden reversal)
- Servo trim applied before constrain (prevented servo damage)
- lastCommandTime updated by WebSocket and UART commands
- Battery percent: int cast before constrain (prevented uint8_t wrap)
- E-stop debounce (50ms), command timeout checks all 4 channels
- Remove PROGMEM/pgm_read_word (no-op on ESP32-S3)
- Web UI: innerHTML to safe DOM methods
- Pin-specific ADC attenuation, integer math in setDrive
- COBS trailing zero removal made robust

ROS2 (10 fixes):
- Add missing nav_msgs dependency (build fix)
- Wire up compute_odometry (was dead code)
- Add mutex for shared UART state (thread safety)
- Fix Gazebo namespace / to match launch files
- Fix WebSocket broadcast with asyncio thread safety
- Waypoint follower: patrol vs single-run mode
- Geofence: publish Bool on /geofence/violation instead of cmd_vel
- Fix rover_autonomy BT directory, rover_bringup data_files

PWA (10 fixes):
- Reconnect dialog re-shown after 10 failed attempts
- IP address validation before WebSocket connection
- Delete stale unreferenced CSS/JS files
- Gamepad stop-on-release, horn mouseleave handler
- E-stop remapped from A button to Back/Select
- Map tiles swap with theme (dark/light)
- Speed slider respects speedLimit setting
- Confirm dialogs for OTA and restart
- Compass 0/360 boundary shortest-path fix

Docs (8 fixes):
- Budget standardized: Phase 1 $102, grand total $2,032
- EA-00: add EA-20/EA-21 references, wheel width 80mm
- EA-05: encoder type optical to hall effect
- EA-09: I2C pins GPIO1/GPIO2, BNO055 not MPU-6050, BME280
- EA-02: note that EA-10 supersedes min turn radius

+++ firmware/esp32/motors.h
@@ -72,13 +72,13 @@ void stopAllMotors() {
 void updateMotors() {
   for (int i = 0; i < 4; i++) {
     if (motorSpeed[i] < motorTarget[i]) {
-      int8_t newSpeed = motorSpeed[i] + RAMP_RATE;
+      int newSpeed = (int)motorSpeed[i] + RAMP_RATE;
       if (newSpeed > motorTarget[i]) newSpeed = motorTarget[i];
-      setMotor(i, newSpeed);
+      setMotor(i, (int8_t)newSpeed);
     } else if (motorSpeed[i] > motorTarget[i]) {
-      int8_t newSpeed = motorSpeed[i] - RAMP_RATE;
+      int newSpeed = (int)motorSpeed[i] - RAMP_RATE;
       if (newSpeed < motorTarget[i]) newSpeed = motorTarget[i];
-      setMotor(i, newSpeed);
+      setMotor(i, (int8_t)newSpeed);
     }
   }
 }
